Run if-vhost0 on during $physdev's pre-up
Running if-vhost0 loads vrouter. Loading vrouter creates vhost0.
Creating vhost0 triggers an upstart event to configure vhost0.

This configures vhost0 as soon as $physdev is configured, which means
it's up early enough for cloud-init to make use of it.

This means cloud-init won't do silly things like the host's name to
hostname.novalocal, etc.

Oh, and it boots faster. Yay.