Optionally disable/enable contrail components
This patch add options to enable/disable specific contrail components if need to
be. One of the situations it would be required is that if you want to separate
some components like analytics to dedicated nodes. Also may be to optionally
disable one component like service monitor.

Added spec tests. Also combined common params for spec tests and merged it to
other params whereever required.